ringbuffer: improve testing, add more benchmarks
