--- internal/utils/ringbuffer/ringbuffer_bench_test.go
@@ -1,14 +1,0 @@
-package ringbuffer
-
-import "testing"
-
-func BenchmarkRingBuffer(b *testing.B) {
-	r := RingBuffer[int]{}
-
-	var val int
-	for b.Loop() {
-		r.PushBack(val)
-		r.PopFront()
-		val++
-	}
-}
